[REFACTOR] 예외 분리 (domain/external)
