+++ smeem-external/src/main/java/com/smeem/external/oauth/apple/exception/AppleException.java
@@ -1,4 +1,4 @@
-package com.smeem.common.exception;
+package com.smeem.external.oauth.apple.exception;
 
 import com.smeem.common.code.failure.FailureCode;
 import lombok.Getter;
